refactor: rename fil på disk til å matche hook-navn
